added login functionality and test

+++ controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogsRouter = require("express").Router();
+const jwt = require("jsonwebtoken");
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const logger = require("../utils/logger");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
 
   response.json(blogs);
 });
@@ -18,13 +21,54 @@ blogsRouter.post("/", async (request, response) => {
   if (body.url === undefined) {
     return response.status(400).end();
   }
-  const blog = new Blog(body);
+  const token = request.token;
+
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
+  const newBlog = {
+    title: body.title,
+    likes: body.likes,
+    url: body.url,
+    author: body.author,
+    user: user._id,
+  };
+
+  const blog = new Blog(newBlog);
 
   const result = await blog.save();
+  user.blogs = user.blogs.concat(result._id);
+  await user.save();
+
   response.status(201).json(result);
 });
 
 blogsRouter.delete("/:id", async (req, res) => {
+  const token = req.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: "token missing or invalid" });
+  }
+  // look for user with token
+  const user = await User.findById(decodedToken.id);
+  if (!user) {
+    return res.status(401).json({ error: "user not found" });
+  }
+  // query blog and grab user, then compare
+  const blog = await Blog.findById(req.params.id);
+  if (!blog) {
+    return res.status(400).json({ error: "blog not found" });
+  }
+  const userId = user._id;
+  if (blog.user.toString() !== userId.toString()) {
+    return res.status(401).json({ error: "invalid permissions for deletion" });
+  }
+  // since everything is good, we can DELET
   await Blog.findByIdAndRemove(req.params.id);
   res.status(204).end();
 });
